application/requirements.txt
sanic>=22.3
sanic-ext
pydantic

sanic-testing
pytest-asyncio